feat: personalize share links
